Consolidate load_payload methods
